add: frontend code
ReactJS application that communicates with Breve API to request short URLs, it returns the response in and redirects in case the user want's to visit the site.

The application use CSS to style the components, and it has input validation as well as error handler in case of API errors responses.
 To configure api url uses: .env file